Fixed week bugs in diagrams

+++ app/compiled/all.js
@@ -927,261 +927,6 @@ function drawImageIOSFix(ctx, img, sx, sy, sw, sh, dx, dy, dw, dh) {
     var vertSquashRatio = detectVerticalSquash(img);
     ctx.drawImage(img, sx, sy, sw, sh, dx, dy, dw, dh / vertSquashRatio);
 }
-/**
- * Mega pixel image rendering library for iOS6 Safari
- *
- * Fixes iOS6 Safari's image file rendering issue for large size image (over mega-pixel),
- * which causes unexpected subsampling when drawing it in canvas.
- * By using this library, you can safely render the image with proper stretching.
- *
- * Copyright (c) 2012 Shinichi Tomita <shinichi.tomita@gmail.com>
- * Released under the MIT license
- */
-(function() {
-
-  /**
-   * Detect subsampling in loaded image.
-   * In iOS, larger images than 2M pixels may be subsampled in rendering.
-   */
-  function detectSubsampling(img) {
-    var iw = img.naturalWidth, ih = img.naturalHeight;
-    if (iw * ih > 1024 * 1024) { // subsampling may happen over megapixel image
-      var canvas = document.createElement('canvas');
-      canvas.width = canvas.height = 1;
-      var ctx = canvas.getContext('2d');
-      ctx.drawImage(img, -iw + 1, 0);
-      // subsampled image becomes half smaller in rendering size.
-      // check alpha channel value to confirm image is covering edge pixel or not.
-      // if alpha value is 0 image is not covering, hence subsampled.
-      return ctx.getImageData(0, 0, 1, 1).data[3] === 0;
-    } else {
-      return false;
-    }
-  }
-
-  /**
-   * Detecting vertical squash in loaded image.
-   * Fixes a bug which squash image vertically while drawing into canvas for some images.
-   */
-  function detectVerticalSquash(img, iw, ih) {
-    var canvas = document.createElement('canvas');
-    canvas.width = 1;
-    canvas.height = ih;
-    var ctx = canvas.getContext('2d');
-    ctx.drawImage(img, 0, 0);
-    var data = ctx.getImageData(0, 0, 1, ih).data;
-    // search image edge pixel position in case it is squashed vertically.
-    var sy = 0;
-    var ey = ih;
-    var py = ih;
-    while (py > sy) {
-      var alpha = data[(py - 1) * 4 + 3];
-      if (alpha === 0) {
-        ey = py;
-      } else {
-        sy = py;
-      }
-      py = (ey + sy) >> 1;
-    }
-    var ratio = (py / ih);
-    return (ratio===0)?1:ratio;
-  }
-
-  /**
-   * Rendering image element (with resizing) and get its data URL
-   */
-  function renderImageToDataURL(img, options, doSquash) {
-    var canvas = document.createElement('canvas');
-    renderImageToCanvas(img, canvas, options, doSquash);
-    return canvas.toDataURL("image/jpeg", options.quality || 0.8);
-  }
-
-  /**
-   * Rendering image element (with resizing) into the canvas element
-   */
-  function renderImageToCanvas(img, canvas, options, doSquash) {
-    var iw = img.naturalWidth, ih = img.naturalHeight;
-    var width = options.width, height = options.height;
-    var ctx = canvas.getContext('2d');
-    ctx.save();
-    transformCoordinate(canvas, width, height, options.orientation);
-    var subsampled = detectSubsampling(img);
-    if (subsampled) {
-      iw /= 2;
-      ih /= 2;
-    }
-    var d = 1024; // size of tiling canvas
-    var tmpCanvas = document.createElement('canvas');
-    tmpCanvas.width = tmpCanvas.height = d;
-    var tmpCtx = tmpCanvas.getContext('2d');
-    var vertSquashRatio = doSquash ? detectVerticalSquash(img, iw, ih) : 1;
-    var dw = Math.ceil(d * width / iw);
-    var dh = Math.ceil(d * height / ih / vertSquashRatio);
-    var sy = 0;
-    var dy = 0;
-    while (sy < ih) {
-      var sx = 0;
-      var dx = 0;
-      while (sx < iw) {
-        tmpCtx.clearRect(0, 0, d, d);
-        tmpCtx.drawImage(img, -sx, -sy);
-        ctx.drawImage(tmpCanvas, 0, 0, d, d, dx, dy, dw, dh);
-        sx += d;
-        dx += dw;
-      }
-      sy += d;
-      dy += dh;
-    }
-    ctx.restore();
-    tmpCanvas = tmpCtx = null;
-  }
-
-  /**
-   * Transform canvas coordination according to specified frame size and orientation
-   * Orientation value is from EXIF tag
-   */
-  function transformCoordinate(canvas, width, height, orientation) {
-    switch (orientation) {
-      case 5:
-      case 6:
-      case 7:
-      case 8:
-        canvas.width = height;
-        canvas.height = width;
-        break;
-      default:
-        canvas.width = width;
-        canvas.height = height;
-    }
-    var ctx = canvas.getContext('2d');
-    switch (orientation) {
-      case 2:
-        // horizontal flip
-        ctx.translate(width, 0);
-        ctx.scale(-1, 1);
-        break;
-      case 3:
-        // 180 rotate left
-        ctx.translate(width, height);
-        ctx.rotate(Math.PI);
-        break;
-      case 4:
-        // vertical flip
-        ctx.translate(0, height);
-        ctx.scale(1, -1);
-        break;
-      case 5:
-        // vertical flip + 90 rotate right
-        ctx.rotate(0.5 * Math.PI);
-        ctx.scale(1, -1);
-        break;
-      case 6:
-        // 90 rotate right
-        ctx.rotate(0.5 * Math.PI);
-        ctx.translate(0, -height);
-        break;
-      case 7:
-        // horizontal flip + 90 rotate right
-        ctx.rotate(0.5 * Math.PI);
-        ctx.translate(width, -height);
-        ctx.scale(-1, 1);
-        break;
-      case 8:
-        // 90 rotate left
-        ctx.rotate(-0.5 * Math.PI);
-        ctx.translate(-width, 0);
-        break;
-      default:
-        break;
-    }
-  }
-
-
-  /**
-   * MegaPixImage class
-   */
-  function MegaPixImage(srcImage) {
-    if (window.Blob && srcImage instanceof Blob) {
-      var img = new Image();
-      var URL = window.URL && window.URL.createObjectURL ? window.URL :
-                window.webkitURL && window.webkitURL.createObjectURL ? window.webkitURL :
-                null;
-      if (!URL) { throw Error("No createObjectURL function found to create blob url"); }
-      img.src = URL.createObjectURL(srcImage);
-      this.blob = srcImage;
-      srcImage = img;
-    }
-    if (!srcImage.naturalWidth && !srcImage.naturalHeight) {
-      var _this = this;
-      srcImage.onload = function() {
-        var listeners = _this.imageLoadListeners;
-        if (listeners) {
-          _this.imageLoadListeners = null;
-          for (var i=0, len=listeners.length; i<len; i++) {
-            listeners[i]();
-          }
-        }
-      };
-      this.imageLoadListeners = [];
-    }
-    this.srcImage = srcImage;
-  }
-
-  /**
-   * Rendering megapix image into specified target element
-   */
-  MegaPixImage.prototype.render = function(target, options) {
-    if (this.imageLoadListeners) {
-      var _this = this;
-      this.imageLoadListeners.push(function() { _this.render(target, options) });
-      return;
-    }
-    options = options || {};
-    var imgWidth = this.srcImage.naturalWidth, imgHeight = this.srcImage.naturalHeight,
-        width = options.width, height = options.height,
-        maxWidth = options.maxWidth, maxHeight = options.maxHeight,
-        doSquash = !this.blob || this.blob.type === 'image/jpeg';
-    if (width && !height) {
-      height = (imgHeight * width / imgWidth) << 0;
-    } else if (height && !width) {
-      width = (imgWidth * height / imgHeight) << 0;
-    } else {
-      width = imgWidth;
-      height = imgHeight;
-    }
-    if (maxWidth && width > maxWidth) {
-      width = maxWidth;
-      height = (imgHeight * width / imgWidth) << 0;
-    }
-    if (maxHeight && height > maxHeight) {
-      height = maxHeight;
-      width = (imgWidth * height / imgHeight) << 0;
-    }
-    var opt = { width : width, height : height };
-    for (var k in options) opt[k] = options[k];
-
-    var tagName = target.tagName.toLowerCase();
-    if (tagName === 'img') {
-      target.src = renderImageToDataURL(this.srcImage, opt, doSquash);
-    } else if (tagName === 'canvas') {
-      renderImageToCanvas(this.srcImage, target, opt, doSquash);
-    }      
-    if (typeof this.onrender === 'function') {
-      this.onrender(target);
-    }
-  };
-
-  /**
-   * Export class to global
-   */
-  if (typeof define === 'function' && define.amd) {
-    define([], function() { return MegaPixImage; }); // for AMD loader
-  } else {
-    this.MegaPixImage = MegaPixImage;
-  }
-
-})();
-
 // moment.js
 // version : 2.0.0
 // author : Tim Wood
@@ -1226,6 +971,9 @@ var GuitarJournalApp = angular.module('GuitarJournalApp', ['ngCookies', 'angles'
       when('/profile', {
         templateUrl: 'profile.html', 
         controller: "ProfileCtrl"
+      }).
+      when('/support', {
+        templateUrl: '/app/support.html'
       }).
       when('/instrument/:id', {
         templateUrl: 'instrument.html', 
@@ -1342,7 +1090,6 @@ function LoginCtrl($scope, $http, $location, $cookies, $cookieStore, $rootScope)
 	$scope.pageSettings.hideNavigation = true;
 	$scope.pageSettings.hideTopNavigation = true;
 	$scope.login = function() {
-//		$http.post("/api/login", {email: $scope.email, password: $scope.password}, $rootScope.httpConfig).success(function(data) {
 		$http.get("/api/login", $rootScope.httpConfig).success(function(data) {
 			if (data._id) {
 				$rootScope.loggedIn = true;
@@ -1445,19 +1192,19 @@ function SessionCtrl($scope, $rootScope, $routeParams, $http, $location, $log, S
 				$scope.showErrorMessage("Error saving session");
 			}
 		);
-
-
 	};
 
 	$scope.delete = function() {
-		Sessions.deleteSession($scope.session._id,
-			function() {
-				$location.path("/sessions/");
-			},
-			function(error) {
-				$scope.showErrorMessage("Could not delete the session.", error);
-			}
-		);
+		if (confirm("Are you sure?")) {
+			Sessions.deleteSession($scope.session._id,
+				function() {
+					$location.path("/sessions/");
+				},
+				function(error) {
+					$scope.showErrorMessage("Could not delete the session.", error);
+				}
+			);
+		}
 	};
 
 	$scope.shareToFacebook = function() {
@@ -1531,13 +1278,15 @@ function GoalCtrl($scope, $routeParams, $http, $location, Goals) {
 	};
 
 	$scope.delete = function() {
-		Goals.deleteGoal($scope.goal._id,
-			function() {
-				$location.path('/goals/');
-			}, 
-			function() {
-				$scope.showErrorMessage("Couldn't delete goal");
-			});
+		if (confirm('Are you sure?')) {
+			Goals.deleteGoal($scope.goal._id,
+				function() {
+					$location.path('/goals/');
+				}, 
+				function() {
+					$scope.showErrorMessage("Couldn't delete goal");
+				});			
+		}
 	};
 }
 
@@ -1615,7 +1364,7 @@ function StatsCtrl($scope, $http, Statistics, Goals, Instruments) {
 	$scope.Math = Math;
 }
 
-function ProfileCtrl($scope, $rootScope, $http, $location, Instruments)
+function ProfileCtrl($scope, $rootScope, $http, $location, Instruments, $window)
 {
 	$scope.Instruments = Instruments;
 	$scope.pageSettings.pageTitle = "Profile";
@@ -1625,15 +1374,9 @@ function ProfileCtrl($scope, $rootScope, $http, $location, Instruments)
 	$scope.pageSettings.hideNavigation = false;
 
 	$scope.logout = function() {
-		fbLogout();
-
-		$http.post('/api/logout', {}, $rootScope.httpConfig)
-			.success(function() {
-				$location.path("/login");
-			})
-			.error(function() {
-				$location.path("/login");
-			});
+		var url = "https://www.facebook.com/logout.php?access_token=" + $rootScope.fbAccessToken +
+			"&confirm=1&next=http://" + $window.location.hostname + "/api/logout";
+		$window.location.href = url;
 	};
 }
 
@@ -1709,91 +1452,7 @@ function InstrumentCtrl($scope, $http, $location, $routeParams, Instruments, $ro
 		    };
 		})(file);
 		reader.readAsDataURL(file);
-
-
-		// MegaPixImage constructor accepts File/Blob object.
-		/*var mpImg = new MegaPixImage(file);
-		var sourceWidth = 0;
-		var sourceHeight = 0;
-		var intermediateSize = 300;
-
-		mpImg.imageLoadListeners.push(function() {
-			// Render resized image into canvas element.
-			var resultCanvas = document.createElement("canvas");
-			mpImg.onrender = function (resizedImage) {
-				var newImage = new Image();
-				newImage.onload = function() {
-					console.log("resizedImage", newImage);
-					console.log("newImage", newImage.naturalWidth, newImage.naturalHeight);
-					var cropCanvas = document.createElement("canvas");
-					var destWidth = 200;
-					var destHeight = 200;
-					var sourceX = Math.round((sourceWidth - intermediateSize) / 2);
-					var sourceY = Math.round((sourceHeight - intermediateSize) / 2);
-					var destX = 0;
-					var destY = 0;
-
-					cropCanvas.width = destWidth;
-					cropCanvas.height = destHeight;
-
-					var context = cropCanvas.getContext('2d');
-					context.drawImage(newImage, sourceX, sourceY, intermediateSize, intermediateSize, destX, destY, destWidth, destHeight);
-
-					var resultData = cropCanvas.toDataURL('image/jpeg', 0.9);
-					console.log("Final image", resultData);
-					$scope.instrument.image = resultData.split(',')[1];					
-				}
-				newImage.src = targetImage.src;
-			}
-			var targetImage = new Image();
-			var largestBound = 0;
-			if (mpImg.srcImage.naturalWidth > mpImg.srcImage.naturalHeight) {
-				largestBound = Math.round(mpImg.srcImage.naturalWidth / (mpImg.srcImage.naturalHeight / intermediateSize));
-				sourceWidth = largestBound;
-				sourceHeight = intermediateSize;
-			}
-			else {
-				largestBound = Math.round(mpImg.srcImage.naturalHeight / (mpImg.srcImage.naturalWidth / intermediateSize));
-				sourceHeight = largestBound;
-				sourceWidth = intermediateSize;
-			}
-			console.log("largestBound", largestBound);
-
-			mpImg.render(targetImage, { maxWidth: largestBound, maxHeight: largestBound });
-		});*/
 	}
-
-	/*
-		$scope.files = imageField.files;
-		var file = $scope.files[0];
-
-		var reader = new FileReader();
-		reader.onload = (function(selectedFile) {
-		    return function(e) {
-				$scope.resizeImage(e.target.result.split(',')[1]);
-		    };
-
-		})(file);
-		reader.readAsDataURL(file);
-	};
-
-	$scope.resizeImage = function(imageFile) {
-		var image = new Image();
-		image.onload=function() {
-    		var maxBoundary = 200;
-   			var r = maxBoundary / Math.max(this.width,this.height);
-			var newWidth = Math.round(this.width * r);
-			var newHeight = Math.round(this.height * r);
-			var canvas = document.createElement("canvas");
-			console.log("Oringal size", this.width, this.height);
-			console.log("New size", newWidth, newHeight);
-			canvas.width = newWidth;
-			canvas.height = newHeight;
-			canvas.getContext("2d").drawImage(this,0,0,newWidth,newHeight);
-			$scope.instrument.image = canvas.toDataURL().split(',')[1];
-		}
-		image.src = "data:image/png;base64," + imageFile;
-	};*/
 
 	$scope.save = function() {
 		Instruments.saveInstrument($scope.instrument,
@@ -2298,22 +1957,32 @@ GuitarJournalApp.factory('Statistics', function($http, $rootScope, $q, $log) {
 			$rootScope.apiStatus.loading++;
 			var url = '/api/statistics/perweek/' + weeks;
 
+			var findWeekData = function(weekDataArray, year, week) {
+				for (var i = 0; i < weekDataArray.length; i++) {
+					if (weekDataArray[i].year == year && weekDataArray[i].week + 1 == week)
+						return weekDataArray[i];
+				}
+
+				return null;
+			};
+
 			$http.get(url)
 				.success(function(data) {
 					if (data && data.length && data.length > 0) {
 						var currentDataIndex = 0;
 
-						for (i = 0; i < weeks; i++) {
-							var currentWeek = moment().subtract(moment.duration(weeks - i, 'weeks')).isoWeek();
+						$log.log("Iterations", weeks);
+						for (var i = 0; i < weeks; i++) {
+							var currentWeek = moment().subtract(moment.duration(weeks - i - 1, 'weeks')).isoWeek();
+							var currentYear = moment().subtract(moment.duration(weeks - i - 1, 'weeks')).year();
+							$log.log("Iteration", i, "Current week", currentWeek, "Current year", currentYear);
+
 							service.sessionsPerWeek.labels.push(currentWeek);
 
-							while (currentDataIndex < (data.length - 1) && data[currentDataIndex].week < currentWeek) {
-								currentDataIndex++;
-							}
-
-							if (data[currentDataIndex].week == currentWeek) {
-								service.sessionsPerWeek.count.push(data[currentDataIndex].count);
-								service.sessionsPerWeek.minutes.push(data[currentDataIndex].minutes);
+							var currentWeekData = findWeekData(data, currentYear, currentWeek);
+							if (currentWeekData != null) {
+								service.sessionsPerWeek.count.push(currentWeekData.count);
+								service.sessionsPerWeek.minutes.push(currentWeekData.minutes);
 
 							}
 							else {
